Добавлен ShapeGroup и методы его отрисовки, получения и установки рамки

+++ Lab7/Lab7/Common/FrameRect.h
@@ -4,6 +4,6 @@
 struct FrameRect
 {
 	Point topLeft;
-	int width;
-	int height;
+	float width;
+	float height;
 };
