+++ Lab7/Lab7/Shape/Rectangle.h
@@ -4,13 +4,13 @@
 class Rectangle : public Shape
 {
 public:
-	Rectangle(const Point& topLeft, int width, int height) noexcept;
+	Rectangle(const Point& topLeft, float width, float height) noexcept;
 
 private:
 	void DrawImpl(ICanvas& canvas) const final;
 	void AdjustShape(const FrameRect& oldFrame, double scaleFactorX, double scaleFactorY) final;
 
 	Point m_topLeft;
-	int m_width;
-	int m_height;
+	float m_width;
+	float m_height;
 };
